pyproject.toml
[project]
name = "tubecast"
version = "0.1.0"
description = "Professional YouTube video uploader with AI-powered metadata generation"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Sunil", email = "sunil@example.com" }]
requires-python = ">=3.10"
keywords = ["youtube", "upload", "video", "cli", "gemini", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "google-api-python-client>=2.86.0",
    "google-auth>=2.18.1",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "oauth2client>=4.1.3",
    "httplib2>=0.22.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "questionary>=2.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
tubecast = "tubecast.cli:app"

[project.urls]
Homepage = "https://github.com/sunil-dhaka/tubecast"
Repository = "https://github.com/sunil-dhaka/tubecast"
Issues = "https://github.com/sunil-dhaka/tubecast/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tubecast"]

[tool.ruff]
line-length = 100
target-version = "py310"
